third_party/blink/web_tests/fast/forms/datetimelocal/input-valueasdate-datetimelocal-expected.txt

Tests for .valueAsDate with <input type=datetimelocal>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS valueAsDateFor("") is null
PASS valueAsDateFor("1969-12-31T12:34:56.789") is null
PASS valueAsDateFor("1970-01-01T00:00:00.000") is null
PASS valueAsDateFor("2009-12-22T11:32:11") is null
PASS successfullyParsed is true

TEST COMPLETE
